Backend/requirements.txt
Flask<=2.0.0
requests
mysql.connector<=2.2.9
sqlalchemy<=1.4.39
pandas>=2.0.0
pymysql>= 1.0.0
scikit-learn
matplotlib==3.7.2
numpy==1.24.3
pymysql
yfinance
seaborn
joblib
flask_cors
flask_session
flask_sqlalchemy
flask_bcrypt
mysql-connector-python